+++ CHANGELOG.md
@@ -1,7 +1,9 @@
 ## next
 
 - Added renderer's name & version into the toolbar
+- Reduced a main bundle size by ~53Kb (12%)
 - Bumped rempl to [`1.0.0-alpha.23`](https://github.com/rempl/rempl/releases/tag/v1.0.0-alpha.20)
+- Made all dependencies as developer only
 
 ## 0.6.2 (May 16, 2022)